requirements.txt.txt
mysql-connector-python
pymongo
python-dotenv